+++ src/main/java/mcjty/lib/gui/RenderGlowEffect.java
@@ -50,10 +50,10 @@ public class RenderGlowEffect {
         int b1 = brightness >> 16 & 65535;
         int b2 = brightness & 65535;
         Quad quad = quads[side];
-        buffer.pos(offs.x + quad.v1.x * mult + offset, offs.y + quad.v1.y * mult + offset, offs.z + quad.v1.z * mult + offset).tex((float) 0, (float) 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
-        buffer.pos(offs.x + quad.v2.x * mult + offset, offs.y + quad.v2.y * mult + offset, offs.z + quad.v2.z * mult + offset).tex((float) 0, (float) 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
-        buffer.pos(offs.x + quad.v3.x * mult + offset, offs.y + quad.v3.y * mult + offset, offs.z + quad.v3.z * mult + offset).tex((float) 1, (float) 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
-        buffer.pos(offs.x + quad.v4.x * mult + offset, offs.y + quad.v4.y * mult + offset, offs.z + quad.v4.z * mult + offset).tex((float) 1, (float) 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(offs.x + quad.v1.x * mult + offset, offs.y + quad.v1.y * mult + offset, offs.z + quad.v1.z * mult + offset).tex(0, 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(offs.x + quad.v2.x * mult + offset, offs.y + quad.v2.y * mult + offset, offs.z + quad.v2.z * mult + offset).tex(0, 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(offs.x + quad.v3.x * mult + offset, offs.y + quad.v3.y * mult + offset, offs.z + quad.v3.z * mult + offset).tex(1, 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(offs.x + quad.v4.x * mult + offset, offs.y + quad.v4.y * mult + offset, offs.z + quad.v4.z * mult + offset).tex(1, 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
     }
 
     public static void addSideFullTexture(BufferBuilder buffer, int side, float mult, float offset) {
@@ -61,10 +61,10 @@ public class RenderGlowEffect {
         int b1 = brightness >> 16 & 65535;
         int b2 = brightness & 65535;
         Quad quad = quads[side];
-        buffer.pos(quad.v1.x * mult + offset, quad.v1.y * mult + offset, quad.v1.z * mult + offset).tex((float) 0, (float) 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
-        buffer.pos(quad.v2.x * mult + offset, quad.v2.y * mult + offset, quad.v2.z * mult + offset).tex((float) 0, (float) 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
-        buffer.pos(quad.v3.x * mult + offset, quad.v3.y * mult + offset, quad.v3.z * mult + offset).tex((float) 1, (float) 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
-        buffer.pos(quad.v4.x * mult + offset, quad.v4.y * mult + offset, quad.v4.z * mult + offset).tex((float) 1, (float) 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(quad.v1.x * mult + offset, quad.v1.y * mult + offset, quad.v1.z * mult + offset).tex(0, 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(quad.v2.x * mult + offset, quad.v2.y * mult + offset, quad.v2.z * mult + offset).tex(0, 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(quad.v3.x * mult + offset, quad.v3.y * mult + offset, quad.v3.z * mult + offset).tex(1, 1).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
+        buffer.pos(quad.v4.x * mult + offset, quad.v4.y * mult + offset, quad.v4.z * mult + offset).tex(1, 0).lightmap(b1, b2).color(255, 255, 255, 128).endVertex();
     }
 
     private static class Vt {
